SimpleTable code refactored
